backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.25.2
pydantic==2.5.0
python-multipart==0.0.6
pyproj==3.6.1
geopy==2.4.1
python-dotenv==1.0.0